backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fives-backend"
version = "0.1.0"
description = "5S Corporate Document Automation System - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "python-docx>=1.1.0",
    "PyMuPDF>=1.24.0",
    "openai>=1.40.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
    "aiofiles>=24.1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
